Create Check Guess method
This commit creates the check_guess method, that take the player's guess
and compare it to the generated code.
The result is a Array, Showing which guesses were correct in both
value and position, which guesses were correct in only value, and which
were wrong. This array is ordered, and to be shown to player, will
need to be treated by the print_hints method.