Create check for victory
If the guess_control array is [1,1,1,1], it means that all guesses
matched the code, in both value and position. This means that the game
ended, and the player won.

The victory? method does this check, and return true or false. This
value is attributed to the instance variable @victory, that gets checked
at the end to display the correct message, either if player Won or Lost